typo-bug: onMultiple color-palette index file
When using multiple languages, the ajax palette-table loader can offer the user a choice, and the page that MasterColorPicker is embedded in can supply a custom js function to handle the human interaction involved in making that choice.  The choice can be at the level of palettes' index, or at the individual palette.  When at the index level, the typo-bug kicked in; now fixed.