major design overhaul
This class has been revamped to be more memory-efficient, user-efficient, flexible, and powerful.  The basic interface is similar; however, there is no longer a need to "register InterfaceElements" (only InterfacePanels now) - i.e. add a group of Event-handlers.   That requirement was fine until the MasterColorPicker "MyPalette" interface started needing to register hundreds of InterfaceElements, and that was becoming a burden.  The original class came from code that was written for when MSIE6 was king and other browsers came along for the ride.  This new version uses a more modern approach to event-handling.