updated; fixed bug in BeezEye-CMYK palette
RGB_Calc was updated and the API was modified a bit.  Changes here reflect that.
There was a bug in the BeezEye-CMYK palette that must have crept in when RGB_Calc was overhauled in the past, likely when it first became a calculator constructor.
RGB_Calc was enhanced to overcome that bug, and this file was modified to use that enhancement as a bug fix.