updated; modernized; small bugs squashed
This release sees the movement from the "alpha" stage to the "beta" stage.
The code was modernized in many ways.
The "x_ColorPicker" class was renamed to "Color_Picker".
The "x_Color" class was renamed to "Color_Picker.Color_SpecCache".
These new names allow better search results if you are working with the code, in conjunction with the names of classes in RGB_Calc.  That file uses the "_Color" postfix for names.
All color-pickers now return an instance of "Color_Picker.Color_SpecCache" or a sub-class.  This class is only used internally by MasterColorPicker as a color is "picked", and is then discarded.  You may add a custom "pick filter" and use this "cache" for whatever you need, and test for it using "instanceof".
All JavaScript-powered HTML output and interfaces were de-localized and conglomerated in the code to allow simple language-packets for multi-lingual support.
The ColorThesaurus formula for comparing colors was modified to give greater weight to the difference in hues.
Minor bug was fixed at the palette-edges in the Simple² Color Picker when the mouse hovers over.
Minor bug was fixed when the "choose color" "button" in the Color-Space Lab has focus and the user presses "Enter"
Other minor bugs.....