updated: enhanced (more?)
I enhanced this file's functionality for another project (the RainbowMaker).
I may have fixed a bug or two, that was a few months ago, and then I got sidetracked with other stuff, so I forget for sure.
Just keeping this project fully updated; the main file "MasterColorPicker2.js" had bugs fixed, and another file ("FormFieldGenie.js") was enhanced with API changes (for the other said project) that required mods to the MasterColorPicker2.js file also, so it was updated at this time.  That may or may not require this updated file, but here it is now anyway.